+++ src/datoso/__init__.py
@@ -3,7 +3,7 @@ __init__.py
 """
 import os
 # Version of the datoso package
-__version__ = "0.2.1.beta1"
+__version__ = "0.2.1"
 __app_name__ = "datoso"
 
 ROOT_FOLDER = os.path.dirname(os.path.abspath(__file__))