+++ WHATSNEW.rst
@@ -1,3 +1,15 @@
+release 0.2.1
+- Seeds are now plugins
+- Removed seed installator/manager
+- Added factory support for actions
+- Internetarchive is now a plugin
+- Cleaned up the code
+- Reorganized the code
+- Dat import of current datdir
+- Added some File and folder helpers
+- Datoso now has a logo (To be changed)
+
+
 release 0.2.0
 - Added --developer to install with git repository instead of zip release
 - Cleaned up the code a bit